Add icons to price history

+++ resources/mysql/create_tables.sql
@@ -20,8 +20,8 @@ CREATE TABLE orderbook (
     num_shares int NOT NULL,
     order_type char(10) NOT NULL,
     price_id int NOT NULL,
-    balance_usd bigint NOT NULL,
-    balance_coin bigint NOT NULL
+    balance_usd decimal(13,2) NOT NULL,
+    balance_coin decimal(13,8) NOT NULL
     CONSTRAINT pk_OrderId PRIMARY KEY(order_id),
     CONSTRAINT fk_AccountId FOREIGN KEY(account_id) REFERENCES accounts(account_id),
     CONSTRAINT fk_PriceId FOREIGN KEY(price_id) REFERENCES prices(price_id)
